Removed edit fragment from media browser
