--- WordPress/src/main/res/menu/media_details.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/menu_edit_media"
-        android:icon="@drawable/ic_pencil_white_24dp"
-        android:title="@string/edit_media"
-        app:showAsAction="ifRoom" />
-
-    <item
-        android:id="@+id/menu_copy_media_url"
-        android:icon="@drawable/ic_copy_white_24dp"
-        android:title="@string/media_details_copy_url"
-        app:showAsAction="ifRoom" />
-
-    <item
-        android:id="@+id/menu_delete"
-        android:icon="@drawable/ic_trash_white_24dp"
-        android:title="@string/delete"
-        app:showAsAction="ifRoom" />
-</menu>
